+++ packages/core/src/services/eth/eth.module.ts
@@ -5,8 +5,16 @@ import { Module } from '@nestd/core'
 import { EthDataService } from './eth-data.service'
 import { ContractService } from './contract.service'
 import { WalletService } from './wallet.service'
+import { EthEventWatcherService } from './events/eth-event-watcher.service'
+import { EthEventHandlerService } from './events/eth-event-handler.service'
 
 @Module({
-  services: [EthDataService, ContractService, WalletService],
+  services: [
+    EthDataService,
+    ContractService,
+    WalletService,
+    EthEventWatcherService,
+    EthEventHandlerService,
+  ],
 })
 export class EthModule {}
