+++ packages/core/src/services/index.ts
@@ -9,15 +9,14 @@ export { JsonRpcService } from './jsonrpc/jsonrpc.service'
 export { DBService } from './db/db.service'
 export { ChainService } from './chain.service'
 export { ProofVerificationService } from './proof/proof-verification.service'
-export { EventWatcherService } from './events/event-watcher.service'
-export { EventHandlerService } from './events/event-handler.service'
 export { OperatorService } from './operator.service'
 export { SyncService } from './sync.service'
 export { WalletService } from './eth/wallet.service'
 export { EthDataService } from './eth/eth-data.service'
+export { EthEventWatcherService } from './eth/events/eth-event-watcher.service'
+export { EthEventHandlerService } from './eth/events/eth-event-handler.service'
 
 /* Modules */
 export { DBModule } from './db/db.module'
 export { EthModule } from './eth/eth.module'
-export { EventModule } from './events/event.module'
 export { ProofModule } from './proof/proof.module'
